Ability to specify chat parameters via HTTP GET querystring I've added ability to override configuration parameters from PS.ChatSDK.Globals.js via querystring parameters. Changes are in following files: •	js/PS.ChatSDK.Globals.js -	changed default value for Calling_User_LastName parameter to blank string (this was hardcoded in js/PS.ChatSDK.js, JoinChat method anyway up to now) -	added AutoConnect, HasFirstName, HasHardMessage, QueryOptionalParameters parameters -	added initGlobalsFromQuerystring definition and execution of initGlobalsFromQuerystring function which parses querystring parameters, overrides ones in JoinChatData and sets new ones (AutoConnect, HasFirstName, HasHardMessage, QueryOptionalParameters) •	js/PS.ChatSDK.js -	changed joinChat method so it does not hardcode Calling_User_LastName and adds chatSDKGlobals.QueryOptionalParameters (initialized by initGlobalsFromQuerystring) to optionalParameterArray •	js/scriptChatAjax.js •	jb/js/scriptChatAjaxOverride.js -	changed top of char.init function to set value for first name and subject (mapped to hardcoded message) fields in case they were provided via querystring parameters -	changed bottom of char.init function to execute connect click handler in case querystring parameter specified it should automatically connect •	README.md -	added list of parameters that can be overridden using querystring parameters in Section 2 -	added example url with querystring parameters in Section 4
Problem 1: Chat not disconnected when user leaves page
When user leaves a page, chat session in progress is sometimes not disconnected. It all depends on the speed of user pressing "leave page" button on page leave warning screen.
The reason for this is that leaveChat function make asynchronous request against server and this request could be terminated before it gets executed if page is unloaded fast.
Changes are in following files:
•	js/PS.ChatSDK.js
o	leaveChat function, to make request synchronous in cases when this is method is called without callback, during page unload (async flag at line 216)

Problem 2: Console warning when user leaves page
There is a warning in browser console about undefined function being invoked when user leaves a page.
The reason for this is that leaveChat is invoked without parameters during page unload but it trys to invoke callback function which is passed to it as a parameter.
Changes are in following files:
•	js/PS.ChatSDK.js
o	leaveChat function, to check for existence of callback function before invoking it (lines 220, 225 and 226)

Problem 3: Chat is disconnected even if user chooses to stay on page
When user tries to leave a page, chat session in progress is disconnected even if user chooses to "stay on page" on page leave warning screen.
The reason for this is that leaveChat is executed inside onbeforeunload, before page leave warning is even displayed to the user.
Changes are in following files:
•	js/scriptChatAjax.js
•	jb/js/scriptChatAjaxOverride.js
-	execute leaveChat inside onunload and not onbeforeunload

Problem 4: Page leave warning screen even without chat session
When user tries to leave a page, page leave warning screen is displayed even if there is no chat session in progress.
The reason for this is that onbeforeunload always returns "".
Changes are in following files:
•	js/scriptChatAjax.js
•	jb/js/scriptChatAjaxOverride.js
-	onbeforeunload to return value based on whether there is a chat in progress or not

Problem 5: Undefined variable console warning
There is a warning in browser console about undefined variable working being used in disconnectButtonOnClick handler and typingMessage handler.
This is probably error made in some refactoring since working variable is seen only inside handlers defined in chat.init.
In addition using this variable and setting it without first checking whether some other handler is already in progress would introduce other errors if executed.
Changes are in following files:
•	js/scriptChatAjax.js
•	jb/js/scriptChatAjaxOverride.js
-	removed usage of this variable in those two methods

Problem 6: Endless status pooling
If you specify wrong queue in configuration or via querystring parameters chat does not get connected but code behind keeps pooling for events indefinitely (untill new chat is connected or page is unloaded).
In addition index2.html displays "Not ConnectedReason: some reason" (notice missing space).
This happens because joinChat succeeds (and sets connection id), first getEvents returns "Not connected" status with some reason and then getEvents gets executed forever.
Changes are in following files:
•	js/PS.ChatSDK.js
o	setting connection id to -1 in case of "Not connected status" (lines 61-62, getEvents function)
•	jb/js/scriptChatAjaxOverride.js
o	added space in onChatStatusReceived

Problem 7: Message typing indicator in Touchpoint stays turned on
Message typing indicator in Touchpoint stays turned on if used enters message and presses OK (or send) within 5 seconds of entering text.
The reason is that send click handler clears onkeyupTimer timer function which sends out typing indicator off message.
Changes are in following files:
•	js/scriptChatAjax.js
•	jb/js/scriptChatAjaxOverride.js
-	removed clearing this timer function in send click handler